shutting up the compiler

+++ src/geoprojectionconverter.cpp
@@ -2388,7 +2388,7 @@ bool GeoProjectionConverter::get_ogc_wkt_from_projection(int& len, char** ogc_wk
       // if not geographic we have a projection
       if ((projection->type != GEO_PROJECTION_LAT_LONG) && (projection->type != GEO_PROJECTION_LONG_LAT))
       {
-        int len = strlen(projection->name);
+        int len = (int)strlen(projection->name);
         char* epsg_name = 0;
         if (len == 0)
         {
@@ -2396,7 +2396,7 @@ bool GeoProjectionConverter::get_ogc_wkt_from_projection(int& len, char** ogc_wk
         }
         else
         {
-          len += strlen(gcs_name) + 16;
+          len += (int)strlen(gcs_name) + 16;
           epsg_name = (char*)malloc(len);
           sprintf(epsg_name, "%s / %s", gcs_name, projection->name);
         }
